精度调整，宽高四舍五入为 0 时使用真实数值

+++ FTMobileSDK/FTSessionReplay/Recorder/SRWireframe/FTSRRecord.h
@@ -61,8 +61,8 @@ NS_ASSUME_NONNULL_BEGIN
 @end
 
 @interface PointerInteractionData : FTSRIncrementalData
-@property (nonatomic, assign) int x;
-@property (nonatomic, assign) int y;
+@property (nonatomic, assign) double x;
+@property (nonatomic, assign) double y;
 @property (nonatomic, assign) int pointerId;
 @property (nonatomic, copy) NSString *pointerEventType;
 @property (nonatomic, copy) NSString *pointerType;
